Refactor ElasticEngine configuration and import
The ElasticEngine's configuration and import statements have been streamlined and reduced in redundancy. The ElasticServers class was removed and replaced with RetrievalServers. The get_config_dir function was also moved to a separate utility module to improve code organization.
